# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API's GIA",
        "description": "API's for transportation management",
        "version": "1.0.0"
    },
    "paths": {
        "/gia-backend/public/api/searchByDni/{dni}": {
            "get": {
                "tags": [
                    "Search"
                ],
                "summary": "Search information by DNI",
                "description": "Search information about a person by their DNI number.",
                "operationId": "e89cd30947f67174513b6eecc3b67c15",
                "parameters": [
                    {
                        "name": "dni",
                        "in": "path",
                        "description": "DNI number of the person to search",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Information found successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "code": {
                                                "type": "integer",
                                                "example": 0
                                            },
                                            "dni": {
                                                "type": "string",
                                                "example": "string"
                                            },
                                            "apepat": {
                                                "type": "string",
                                                "example": "string"
                                            },
                                            "apemat": {
                                                "type": "string",
                                                "example": "string"
                                            },
                                            "apcas": {
                                                "type": "string",
                                                "example": ""
                                            },
                                            "nombres": {
                                                "type": "string",
                                                "example": "string"
                                            },
                                            "fecnac": {
                                                "type": "string",
                                                "format": "date"
                                            },
                                            "ubigeo": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/gia-backend/public/api/searchByRuc/{ruc}": {
            "get": {
                "tags": [
                    "Search"
                ],
                "summary": "Search information by RUC",
                "description": "Search information about a person by their RUC number.",
                "operationId": "5a3423112cb2e5db0ad106daebce9185",
                "parameters": [
                    {
                        "name": "ruc",
                        "in": "path",
                        "description": "RUC number of the person to search",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Information found successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "code": {
                                                "type": "integer",
                                                "example": 0
                                            },
                                            "RUC": {
                                                "type": "string",
                                                "example": "string"
                                            },
                                            "RazonSocial": {
                                                "type": "string",
                                                "example": "string"
                                            },
                                            "Direccion": {
                                                "type": "string",
                                                "example": "string"
                                            },
                                            "Tipo": {
                                                "type": "string"
                                            },
                                            "Inscripcion": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}